pyproject.toml
[project]
name = "search-customizations"
version = "0.1.0"
description = "Customizing the Azure Search AI using Azure Functions"
authors = [
    {name = "Okwudili Ezeme"}
]
license = {text = "BSD 3-Clause"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-search-documents (>=11.5.2,<12.0.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",  # Added new dependency
    "azure-functions (>=1.21.3,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
